+++ templates/single-tapestry.php
@@ -4,49 +4,6 @@
  Template Name: Tapestry Page Template
  */
 
-/**
- * Register Script with Nonce
- * 
- * @return Object null
- */
-function addNonceToScript()
-{
-    $params = array(
-        'nonce'  => wp_create_nonce('wp_rest')
-    );
-
-    wp_register_script(
-        'wp_tapestry_script',
-        plugin_dir_url(__FILE__) . 'tapestry-d3/tapestry.js',
-        array('jquery'),
-        null,
-        true
-    );
-    wp_localize_script('wp_tapestry_script', 'wpApiSettings', $params);
-    wp_enqueue_script('wp_tapestry_script');
-
-    wp_register_script(
-        'wp_tapestry_functions_script',
-        plugin_dir_url(__FILE__) . 'tapestry-d3/tapestry-functions.js',
-        array('jquery', 'wp_tapestry_script'),
-        null,
-        true
-    );
-    wp_enqueue_script('wp_tapestry_functions_script');
-
-    wp_add_inline_script( 'wp_tapestry_script', "
-        var thisTapestryTool = new tapestryTool({
-            'containerId': 'tapestry-container',
-            'apiUrl': '". get_rest_url(null, 'tapestry-tool/v1') ."',
-            'wpUserId': '". apply_filters('determine_current_user', false) ."',
-            'wpPostId': '". get_the_ID() ."',
-            'wpIsAdmin': '". current_user_can('administrator') ."',
-            'addNodeModalUrl': '". plugin_dir_url( __FILE__ ) ."modal-add-node.html',
-        });
-    " );
-}
-add_action('wp_enqueue_scripts', 'addNonceToScript');
-
 function enqueue_vue_app_build()
 {
     // register the Vue build script.
@@ -80,6 +37,50 @@ function enqueue_vue_app_build()
     wp_enqueue_script('tapestry_d3_vue');
 }
 add_action('wp_enqueue_scripts', 'enqueue_vue_app_build');
+
+/**
+ * Register Script with Nonce
+ * 
+ * @return Object null
+ */
+function addNonceToScript()
+{
+    $params = array(
+        'nonce'  => wp_create_nonce('wp_rest')
+    );
+
+    wp_register_script(
+        'wp_tapestry_functions_script',
+        plugin_dir_url(__FILE__) . 'libs/tapestry-functions.js',
+        array('jquery', 'wp_tapestry_script'),
+        null,
+        true
+    );
+    wp_enqueue_script('wp_tapestry_functions_script');
+
+    wp_register_script(
+        'wp_tapestry_script',
+        plugin_dir_url(__FILE__) . 'libs/tapestry.js',
+        array('jquery'),
+        null,
+        true
+    );
+    wp_localize_script('wp_tapestry_script', 'wpApiSettings', $params);
+    wp_enqueue_script('wp_tapestry_script');
+
+
+    wp_add_inline_script('wp_tapestry_script', "
+        var thisTapestryTool = new tapestryTool({
+            'containerId': 'tapestry',
+            'apiUrl': '" . get_rest_url(null, 'tapestry-tool/v1') . "',
+            'wpUserId': '" . apply_filters('determine_current_user', false) . "',
+            'wpPostId': '" . get_the_ID() . "',
+            'wpIsAdmin': '" . current_user_can('administrator') . "',
+            'addNodeModalUrl': '" . plugin_dir_url(__FILE__) . "modal-add-node.html',
+        });
+    ");
+}
+add_action('wp_enqueue_scripts', 'addNonceToScript');
 
 get_header(); ?>
 
@@ -117,23 +118,30 @@ get_header(); ?>
             // thisTapestryTool.setDataset({'abc':'123'});
             // thisTapestryTool.redraw(false);
 
+            var wpPostId = "<?php echo get_the_ID(); ?>";
+            var apiUrl = "<?php echo get_rest_url(null, 'tapestry-tool/v1'); ?>";
+
             // Capture click events anywhere inside or outside tapestry
-            $(document).ready(function(){
+            $(document).ready(function() {
                 document.body.addEventListener('click', function(event) {
                     var x = event.clientX + $(window).scrollLeft();
                     var y = event.clientY + $(window).scrollTop();
-                    recordAnalyticsEvent('user', 'click', 'screen', null, {'x': x, 'y': y});
+                    recordAnalyticsEvent('user', 'click', 'screen', null, {
+                        'x': x,
+                        'y': y
+                    });
                 }, true);
 
-                document.getElementById('tapestry-container').addEventListener('click', function(event) {
+                document.getElementById('tapestry').addEventListener('click', function(event) {
                     var x = event.clientX + $(window).scrollLeft();
                     var y = event.clientY + $(window).scrollTop();
-                    recordAnalyticsEvent('user', 'click', 'tapestry', null, {'x': x, 'y': y});
+                    recordAnalyticsEvent('user', 'click', 'tapestry', null, {
+                        'x': x,
+                        'y': y
+                    });
                 }, true);
             });
         </script>
-
-        <script src="<?php echo plugin_dir_url(__FILE__) ?>libs/tapestry.js" type="application/javascript"></script>
     </main><!-- #main -->
 </div><!-- #primary -->
